requirements.txt
Flask==2.0.3
torch==1.10.0
torchvision==0.11.1
spacy==3.1.0
pillow==8.3.2
tqdm==4.62.3
